@@ -9,11 +9,11 @@ import net.badata.protobuf.example.server.ProtobufServer;
  */
 public class Main {
 
-	public static void main(final String[] args) throws Exception {
-		if(args != null && args.length > 0 && "server".equals(args[0])) {
-			new ProtobufServer().start();
-		} else {
-			new ClientWindow();
-		}
-	}
+    public static void main(final String[] args) throws Exception {
+        if (args != null && args.length > 0 && "server".equals(args[0])) {
+            new ProtobufServer().start();
+        } else {
+            new ClientWindow();
+        }
+    }
 }
