@@ -9,14 +9,13 @@ import net.badata.protobuf.example.proto.BookState;
  */
 public class BooleanEnumConverter implements TypeConverter<Boolean, BookState> {
 
+    @Override
+    public Boolean toDomainValue(final Object object) {
+        return BookState.AVAILABLE == object;
+    }
 
-	@Override
-	public Boolean toDomainValue(final Object object) {
-		return BookState.AVAILABLE == object;
-	}
-
-	@Override
-	public BookState toProtobufValue(final Object object) {
-		return ((Boolean) object).booleanValue() ? BookState.AVAILABLE : BookState.UNAVAILABLE;
-	}
+    @Override
+    public BookState toProtobufValue(final Object object) {
+        return (Boolean) object ? BookState.AVAILABLE : BookState.UNAVAILABLE;
+    }
 }
