@@ -1,5 +1,7 @@
 package net.badata.protobuf.example.client.domain;
 
+import lombok.Getter;
+import lombok.Setter;
 import net.badata.protobuf.converter.annotation.ProtoClass;
 import net.badata.protobuf.converter.annotation.ProtoField;
 import net.badata.protobuf.example.proto.User;
@@ -8,27 +10,14 @@ import net.badata.protobuf.example.proto.User;
  * @author jsjem
  * @author Roman Gushel
  */
+@Getter
 @ProtoClass(User.class)
+@Setter
 public class Reader {
 
-	@ProtoField
-	private String name;
-	@ProtoField
-	private String password;
+    @ProtoField
+    private String name;
 
-	public String getName() {
-		return name;
-	}
-
-	public void setName(final String name) {
-		this.name = name;
-	}
-
-	public String getPassword() {
-		return password;
-	}
-
-	public void setPassword(final String password) {
-		this.password = password;
-	}
+    @ProtoField
+    private String password;
 }
