@@ -1,14 +1,13 @@
 package net.badata.protobuf.example.server;
 
-import net.badata.protobuf.example.server.domain.LibraryBook;
-import net.badata.protobuf.example.server.domain.Reader;
-
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
+import net.badata.protobuf.example.server.domain.LibraryBook;
+import net.badata.protobuf.example.server.domain.Reader;
 
 /**
  * @author jsjem
@@ -16,43 +15,44 @@ import java.util.Set;
  */
 public class MemoryDatabase {
 
-	private final Set<LibraryBook> books = new HashSet<>();
-	private final Map<String, Reader> users = new HashMap<>();
+    private final Set<LibraryBook> books = new HashSet<>();
+    private final Map<String, Reader> users = new HashMap<>();
 
+    public void addUser(final Reader user, final String token) {
+        users.put(token, user);
+    }
 
-	public void addUser(final Reader user, final String token) {
-		users.put(token, user);
-	}
+    public Reader findUser(final String token) {
+        return users.get(token);
+    }
 
-	public Reader findUser(final String token) {
-		return users.get(token);
-	}
+    public void putBook(final LibraryBook book) {
+        if (!books.add(book) && books.remove(book)) {
+            books.add(book);
+        }
+    }
 
-	public void putBook(final LibraryBook book) {
-		if(!books.add(book) && books.remove(book)) {
-			books.add(book);
-		}
-	}
+    public boolean giveBook(final LibraryBook book, final Reader user) {
+        for (LibraryBook storedBook : books) {
+            if (storedBook.equals(book)) {
+                return makeUnavailable(storedBook, user);
+            }
+        }
 
-	public boolean giveBook(final LibraryBook book, final Reader user) {
-		for (LibraryBook storedBook : books) {
-			if (storedBook.equals(book)) {
-				return makeUnavailable(storedBook, user);
-			}
-		}
-		return false;
-	}
+        return false;
+    }
 
-	private boolean makeUnavailable(final LibraryBook storedBook, final Reader user) {
-		if (storedBook.isAvailable()) {
-			storedBook.setAvailable(false);
-			storedBook.setOwner(user);
-			return true;
-		}
-		return false;
-	}
+    private boolean makeUnavailable(final LibraryBook storedBook, final Reader user) {
+        if (storedBook.isAvailable()) {
+            storedBook.setAvailable(false);
+            storedBook.setOwner(user);
 
-	public List<LibraryBook> getAllBooks() {
-		return new ArrayList<LibraryBook>(books);
-	}
+            return true;
+        }
+        return false;
+    }
+
+    public List<LibraryBook> getAllBooks() {
+        return new ArrayList<>(books);
+    }
 }
