@@ -1,6 +1,5 @@
 package net.badata.protobuf.converter.type;
 
-
 /**
  * Implementation of {@link net.badata.protobuf.converter.type.TypeConverter TypeConverter} that is
  * applied by default.
@@ -12,19 +11,19 @@ package net.badata.protobuf.converter.type;
  */
 public class DefaultConverterImpl implements TypeConverter<Object, Object> {
 
-	/**
-	 * {@inheritDoc}
-	 */
-	@Override
-	public Object toDomainValue(final Object instance) {
-		return instance;
-	}
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public Object toDomainValue(final Object instance) {
+        return instance;
+    }
 
-	/**
-	 * {@inheritDoc}
-	 */
-	@Override
-	public Object toProtobufValue(final Object instance) {
-		return instance;
-	}
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public Object toProtobufValue(final Object instance) {
+        return instance;
+    }
 }
