@@ -11,37 +11,37 @@ import net.badata.protobuf.converter.resolver.FieldResolver;
  */
 public abstract class AbstractWriter {
 
-	private final Object destination;
+    private final Object destination;
 
-	/**
-	 * Constructor.
-	 *
-	 * @param destination Destination object to which data has to be written.
-	 */
-	protected AbstractWriter(final Object destination) {
-		this.destination = destination;
-	}
+    /**
+     * Constructor.
+     *
+     * @param destination Destination object to which data has to be written.
+     */
+    protected AbstractWriter(final Object destination) {
+        this.destination = destination;
+    }
 
-	/**
-	 * Write data to the destination object.
-	 *
-	 * @param fieldResolver Field resolver which value has to be written.
-	 * @param value         Value to write.
-	 * @throws WriteException throws when some data writing errors happens. E.g.: destination setters is inaccessible,
-	 *                        etc.
-	 */
-	public void write(final FieldResolver fieldResolver, final Object value) throws WriteException {
-		write(destination, fieldResolver, value);
-	}
+    /**
+     * Write data to the destination object.
+     *
+     * @param fieldResolver Field resolver which value has to be written.
+     * @param value         Value to write.
+     * @throws WriteException throws when some data writing errors happens. E.g.: destination setters is inaccessible,
+     *                        etc.
+     */
+    public void write(final FieldResolver fieldResolver, final Object value) throws WriteException {
+        write(destination, fieldResolver, value);
+    }
 
-	/**
-	 * Write data to the destination object.
-	 *
-	 * @param destination   Instance to which data has to be written.
-	 * @param fieldResolver Field resolver which value has to be written.
-	 * @param value         Value to write.
-	 * @throws WriteException throws when some data writing errors happens.
-	 */
-	protected abstract void write(final Object destination, final FieldResolver fieldResolver, final Object value)
-			throws WriteException;
+    /**
+     * Write data to the destination object.
+     *
+     * @param destination   Instance to which data has to be written.
+     * @param fieldResolver Field resolver which value has to be written.
+     * @param value         Value to write.
+     * @throws WriteException throws when some data writing errors happens.
+     */
+    protected abstract void write(final Object destination, final FieldResolver fieldResolver, final Object value)
+        throws WriteException;
 }
