@@ -1,5 +1,8 @@
 package net.badata.protobuf.converter.domain;
 
+import lombok.Data;
+import lombok.Getter;
+import lombok.Setter;
 import net.badata.protobuf.converter.annotation.ProtoClass;
 import net.badata.protobuf.converter.annotation.ProtoField;
 import net.badata.protobuf.converter.proto.DomainInheritanceProto;
@@ -9,50 +12,27 @@ import net.badata.protobuf.converter.proto.DomainInheritanceProto;
  * @author Roman Gushel
  */
 public class DomainInheritanceDomain {
-	@ProtoClass(DomainInheritanceProto.Test.class)
-	public static class Test extends InheritedByTest {
-		@ProtoField
-		private Long ownLong;
-		@ProtoField
-		private double ownDouble;
 
-		public Long getOwnLong() {
-			return ownLong;
-		}
+    @Getter
+    @ProtoClass(DomainInheritanceProto.Test.class)
+    @Setter
+    public static class Test extends InheritedByTest {
 
-		public void setOwnLong(Long ownLong) {
-			this.ownLong = ownLong;
-		}
+        @ProtoField
+        private Long ownLong;
 
-		public double getOwnDouble() {
-			return ownDouble;
-		}
+        @ProtoField
+        private double ownDouble;
+    }
 
-		public void setOwnDouble(double ownDouble) {
-			this.ownDouble = ownDouble;
-		}
-	}
+    @Getter
+    @Setter
+    public static class InheritedByTest {
 
-	public static class InheritedByTest {
-		@ProtoField
-		private int inheritedInt;
-		@ProtoField
-		private Float inheritedFloat;
+        @ProtoField
+        private int inheritedInt;
 
-		public int getInheritedInt() {
-			return inheritedInt;
-		}
-
-		public void setInheritedInt(int inheritedInt) {
-			this.inheritedInt = inheritedInt;
-		}
-
-		public Float getInheritedFloat() {
-			return inheritedFloat;
-		}
-
-		public void setInheritedFloat(Float inheritedFloat) {
-			this.inheritedFloat = inheritedFloat;
-		}
-	}
+        @ProtoField
+        private Float inheritedFloat;
+    }
 }
