@@ -1,10 +1,11 @@
 package net.badata.protobuf.converter.domain;
 
+import java.util.List;
+import lombok.Getter;
+import lombok.Setter;
 import net.badata.protobuf.converter.annotation.ProtoClass;
 import net.badata.protobuf.converter.annotation.ProtoField;
 import net.badata.protobuf.converter.proto.MappingProto;
-
-import java.util.List;
 
 /**
  * @author jsjem
@@ -12,184 +13,77 @@ import java.util.List;
  */
 public class MappingDomain {
 
-	@ProtoClass(MappingProto.MappingTest.class)
-	public static class Test {
+    @Getter
+    @ProtoClass(MappingProto.MappingTest.class)
+    @Setter
+    public static class Test {
 
-		@ProtoField
-		private Long longValue;
-		@ProtoField
-		private Integer intValue;
-		@ProtoField
-		private Float floatValue;
-		@ProtoField
-		private Double doubleValue;
-		@ProtoField(name = "booleanValue")
-		private Boolean boolValue;
-		@ProtoField
-		private String stringValue;
-		@ProtoField
-		private NestedTest nestedValue;
-		@ProtoField(name = "stringListValue")
-		private List<String> simpleListValue;
-		@ProtoField
-		private List<NestedTest> nestedListValue;
+        @ProtoField
+        private Long longValue;
 
+        @ProtoField
+        private Integer intValue;
 
+        @ProtoField
+        private Float floatValue;
 
-		public Long getLongValue() {
-			return longValue;
-		}
+        @ProtoField
+        private Double doubleValue;
 
-		public void setLongValue(final Long longValue) {
-			this.longValue = longValue;
-		}
+        @ProtoField(name = "booleanValue")
+        private Boolean boolValue;
 
-		public Integer getIntValue() {
-			return intValue;
-		}
+        @ProtoField
+        private String stringValue;
 
-		public void setIntValue(final Integer intValue) {
-			this.intValue = intValue;
-		}
+        @ProtoField
+        private NestedTest nestedValue;
 
-		public Float getFloatValue() {
-			return floatValue;
-		}
+        @ProtoField(name = "stringListValue")
+        private List<String> simpleListValue;
 
-		public void setFloatValue(final Float floatValue) {
-			this.floatValue = floatValue;
-		}
+        @ProtoField
+        private List<NestedTest> nestedListValue;
+    }
 
-		public Double getDoubleValue() {
-			return doubleValue;
-		}
+    @Getter
+    @ProtoClass(MappingProto.NestedTest.class)
+    @Setter
+    public static class NestedTest {
 
-		public void setDoubleValue(final Double doubleValue) {
-			this.doubleValue = doubleValue;
-		}
+        @ProtoField
+        private String stringValue;
+    }
 
-		public Boolean getBoolValue() {
-			return boolValue;
-		}
+    @Getter
+    @ProtoClass(MappingProto.MappingTest.class)
+    @Setter
+    public static class PrimitiveTest {
 
-		public void setBoolValue(final Boolean boolValue) {
-			this.boolValue = boolValue;
-		}
+        @ProtoField
+        private long longValue;
 
-		public String getStringValue() {
-			return stringValue;
-		}
+        @ProtoField
+        private int intValue;
 
-		public void setStringValue(final String stringValue) {
-			this.stringValue = stringValue;
-		}
+        @ProtoField
+        private float floatValue;
 
-		public NestedTest getNestedValue() {
-			return nestedValue;
-		}
+        @ProtoField
+        private double doubleValue;
 
-		public void setNestedValue(final NestedTest nestedValue) {
-			this.nestedValue = nestedValue;
-		}
+        @ProtoField
+        private boolean booleanValue;
+    }
 
-		public List<String> getSimpleListValue() {
-			return simpleListValue;
-		}
+    @ProtoClass(MappingProto.MappingTest.class)
+    public static class InaccessibleTest {
 
-		public void setSimpleListValue(final List<String> simpleListValue) {
-			this.simpleListValue = simpleListValue;
-		}
+        @ProtoField
+        private Object inaccessibleField;
 
-		public List<NestedTest> getNestedListValue() {
-			return nestedListValue;
-		}
-
-		public void setNestedListValue(final List<NestedTest> nestedListValue) {
-			this.nestedListValue = nestedListValue;
-		}
-	}
-
-	@ProtoClass(MappingProto.NestedTest.class)
-	public static class NestedTest {
-
-		@ProtoField
-		private String stringValue;
-
-		public String getStringValue() {
-			return stringValue;
-		}
-
-		public void setStringValue(final String stringValue) {
-			this.stringValue = stringValue;
-		}
-	}
-
-	@ProtoClass(MappingProto.MappingTest.class)
-	public static class PrimitiveTest {
-
-		@ProtoField
-		private long longValue;
-		@ProtoField
-		private int intValue;
-		@ProtoField
-		private float floatValue;
-		@ProtoField
-		private double doubleValue;
-		@ProtoField
-		private boolean booleanValue;
-
-		public long getLongValue() {
-			return longValue;
-		}
-
-		public void setLongValue(final long longValue) {
-			this.longValue = longValue;
-		}
-
-		public int getIntValue() {
-			return intValue;
-		}
-
-		public void setIntValue(final int intValue) {
-			this.intValue = intValue;
-		}
-
-		public float getFloatValue() {
-			return floatValue;
-		}
-
-		public void setFloatValue(final float floatValue) {
-			this.floatValue = floatValue;
-		}
-
-		public double getDoubleValue() {
-			return doubleValue;
-		}
-
-		public void setDoubleValue(final double doubleValue) {
-			this.doubleValue = doubleValue;
-		}
-
-		public boolean isBooleanValue() {
-			return booleanValue;
-		}
-
-		public void setBooleanValue(final boolean booleanValue) {
-			this.booleanValue = booleanValue;
-		}
-	}
-
-	@ProtoClass(MappingProto.MappingTest.class)
-	public static class InaccessibleTest {
-
-		@ProtoField
-		private Object inaccessibleField;
-		@ProtoField
-		private Object protectedGetterField;
-
-		protected Object getProtectedGetterField() {
-			return protectedGetterField;
-		}
-	}
-
+        @Getter
+        @ProtoField
+        private Object protectedGetterField;
+    }
 }
