@@ -4,774 +4,795 @@
 package net.badata.protobuf.converter.proto;
 
 public final class ResolverProto {
-  private ResolverProto() {}
-  public static void registerAllExtensions(
-      com.google.protobuf.ExtensionRegistryLite registry) {
-  }
 
-  public static void registerAllExtensions(
-      com.google.protobuf.ExtensionRegistry registry) {
-    registerAllExtensions(
-        (com.google.protobuf.ExtensionRegistryLite) registry);
-  }
-  public interface ResolverTestOrBuilder extends
-      // @@protoc_insertion_point(interface_extends:net.badata.protobuf.converter.proto.ResolverTest)
-      com.google.protobuf.MessageOrBuilder {
+    private ResolverProto() {}
 
-    /**
-     * <code>repeated string stringListValue = 1;</code>
-     */
-    java.util.List<java.lang.String>
-        getStringListValueList();
-    /**
-     * <code>repeated string stringListValue = 1;</code>
-     */
-    int getStringListValueCount();
-    /**
-     * <code>repeated string stringListValue = 1;</code>
-     */
-    java.lang.String getStringListValue(int index);
-    /**
-     * <code>repeated string stringListValue = 1;</code>
-     */
-    com.google.protobuf.ByteString
-        getStringListValueBytes(int index);
+    public static void registerAllExtensions(com.google.protobuf.ExtensionRegistryLite registry) {}
 
-    /**
-     * <code>string delimitedStringValue = 3;</code>
-     */
-    java.lang.String getDelimitedStringValue();
-    /**
-     * <code>string delimitedStringValue = 3;</code>
-     */
-    com.google.protobuf.ByteString
-        getDelimitedStringValueBytes();
-  }
-  /**
-   * Protobuf type {@code net.badata.protobuf.converter.proto.ResolverTest}
-   */
-  public  static final class ResolverTest extends
-      com.google.protobuf.GeneratedMessageV3 implements
-      // @@protoc_insertion_point(message_implements:net.badata.protobuf.converter.proto.ResolverTest)
-      ResolverTestOrBuilder {
-  private static final long serialVersionUID = 0L;
-    // Use ResolverTest.newBuilder() to construct.
-    private ResolverTest(com.google.protobuf.GeneratedMessageV3.Builder<?> builder) {
-      super(builder);
-    }
-    private ResolverTest() {
-      stringListValue_ = com.google.protobuf.LazyStringArrayList.EMPTY;
-      delimitedStringValue_ = "";
+    public static void registerAllExtensions(com.google.protobuf.ExtensionRegistry registry) {
+        registerAllExtensions((com.google.protobuf.ExtensionRegistryLite) registry);
     }
 
-    @java.lang.Override
-    public final com.google.protobuf.UnknownFieldSet
-    getUnknownFields() {
-      return this.unknownFields;
-    }
-    private ResolverTest(
-        com.google.protobuf.CodedInputStream input,
-        com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-        throws com.google.protobuf.InvalidProtocolBufferException {
-      this();
-      if (extensionRegistry == null) {
-        throw new java.lang.NullPointerException();
-      }
-      int mutable_bitField0_ = 0;
-      com.google.protobuf.UnknownFieldSet.Builder unknownFields =
-          com.google.protobuf.UnknownFieldSet.newBuilder();
-      try {
-        boolean done = false;
-        while (!done) {
-          int tag = input.readTag();
-          switch (tag) {
-            case 0:
-              done = true;
-              break;
-            default: {
-              if (!parseUnknownFieldProto3(
-                  input, unknownFields, extensionRegistry, tag)) {
-                done = true;
-              }
-              break;
-            }
-            case 10: {
-              java.lang.String s = input.readStringRequireUtf8();
-              if (!((mutable_bitField0_ & 0x00000001) == 0x00000001)) {
-                stringListValue_ = new com.google.protobuf.LazyStringArrayList();
-                mutable_bitField0_ |= 0x00000001;
-              }
-              stringListValue_.add(s);
-              break;
-            }
-            case 26: {
-              java.lang.String s = input.readStringRequireUtf8();
+    public interface ResolverTestOrBuilder
+        extends
+            // @@protoc_insertion_point(interface_extends:net.badata.protobuf.converter.proto.ResolverTest)
+            com.google.protobuf.MessageOrBuilder {
+        /**
+         * <code>repeated string stringListValue = 1;</code>
+         */
+        java.util.List<java.lang.String> getStringListValueList();
+        /**
+         * <code>repeated string stringListValue = 1;</code>
+         */
+        int getStringListValueCount();
+        /**
+         * <code>repeated string stringListValue = 1;</code>
+         */
+        java.lang.String getStringListValue(int index);
+        /**
+         * <code>repeated string stringListValue = 1;</code>
+         */
+        com.google.protobuf.ByteString getStringListValueBytes(int index);
 
-              delimitedStringValue_ = s;
-              break;
-            }
-          }
-        }
-      } catch (com.google.protobuf.InvalidProtocolBufferException e) {
-        throw e.setUnfinishedMessage(this);
-      } catch (java.io.IOException e) {
-        throw new com.google.protobuf.InvalidProtocolBufferException(
-            e).setUnfinishedMessage(this);
-      } finally {
-        if (((mutable_bitField0_ & 0x00000001) == 0x00000001)) {
-          stringListValue_ = stringListValue_.getUnmodifiableView();
-        }
-        this.unknownFields = unknownFields.build();
-        makeExtensionsImmutable();
-      }
-    }
-    public static final com.google.protobuf.Descriptors.Descriptor
-        getDescriptor() {
-      return net.badata.protobuf.converter.proto.ResolverProto.internal_static_net_badata_protobuf_converter_proto_ResolverTest_descriptor;
+        /**
+         * <code>string delimitedStringValue = 3;</code>
+         */
+        java.lang.String getDelimitedStringValue();
+        /**
+         * <code>string delimitedStringValue = 3;</code>
+         */
+        com.google.protobuf.ByteString getDelimitedStringValueBytes();
     }
 
-    protected com.google.protobuf.GeneratedMessageV3.FieldAccessorTable
-        internalGetFieldAccessorTable() {
-      return net.badata.protobuf.converter.proto.ResolverProto.internal_static_net_badata_protobuf_converter_proto_ResolverTest_fieldAccessorTable
-          .ensureFieldAccessorsInitialized(
-              net.badata.protobuf.converter.proto.ResolverProto.ResolverTest.class, net.badata.protobuf.converter.proto.ResolverProto.ResolverTest.Builder.class);
-    }
-
-    private int bitField0_;
-    public static final int STRINGLISTVALUE_FIELD_NUMBER = 1;
-    private com.google.protobuf.LazyStringList stringListValue_;
-    /**
-     * <code>repeated string stringListValue = 1;</code>
-     */
-    public com.google.protobuf.ProtocolStringList
-        getStringListValueList() {
-      return stringListValue_;
-    }
-    /**
-     * <code>repeated string stringListValue = 1;</code>
-     */
-    public int getStringListValueCount() {
-      return stringListValue_.size();
-    }
-    /**
-     * <code>repeated string stringListValue = 1;</code>
-     */
-    public java.lang.String getStringListValue(int index) {
-      return stringListValue_.get(index);
-    }
-    /**
-     * <code>repeated string stringListValue = 1;</code>
-     */
-    public com.google.protobuf.ByteString
-        getStringListValueBytes(int index) {
-      return stringListValue_.getByteString(index);
-    }
-
-    public static final int DELIMITEDSTRINGVALUE_FIELD_NUMBER = 3;
-    private volatile java.lang.Object delimitedStringValue_;
-    /**
-     * <code>string delimitedStringValue = 3;</code>
-     */
-    public java.lang.String getDelimitedStringValue() {
-      java.lang.Object ref = delimitedStringValue_;
-      if (ref instanceof java.lang.String) {
-        return (java.lang.String) ref;
-      } else {
-        com.google.protobuf.ByteString bs = 
-            (com.google.protobuf.ByteString) ref;
-        java.lang.String s = bs.toStringUtf8();
-        delimitedStringValue_ = s;
-        return s;
-      }
-    }
-    /**
-     * <code>string delimitedStringValue = 3;</code>
-     */
-    public com.google.protobuf.ByteString
-        getDelimitedStringValueBytes() {
-      java.lang.Object ref = delimitedStringValue_;
-      if (ref instanceof java.lang.String) {
-        com.google.protobuf.ByteString b = 
-            com.google.protobuf.ByteString.copyFromUtf8(
-                (java.lang.String) ref);
-        delimitedStringValue_ = b;
-        return b;
-      } else {
-        return (com.google.protobuf.ByteString) ref;
-      }
-    }
-
-    private byte memoizedIsInitialized = -1;
-    public final boolean isInitialized() {
-      byte isInitialized = memoizedIsInitialized;
-      if (isInitialized == 1) return true;
-      if (isInitialized == 0) return false;
-
-      memoizedIsInitialized = 1;
-      return true;
-    }
-
-    public void writeTo(com.google.protobuf.CodedOutputStream output)
-                        throws java.io.IOException {
-      for (int i = 0; i < stringListValue_.size(); i++) {
-        com.google.protobuf.GeneratedMessageV3.writeString(output, 1, stringListValue_.getRaw(i));
-      }
-      if (!getDelimitedStringValueBytes().isEmpty()) {
-        com.google.protobuf.GeneratedMessageV3.writeString(output, 3, delimitedStringValue_);
-      }
-      unknownFields.writeTo(output);
-    }
-
-    public int getSerializedSize() {
-      int size = memoizedSize;
-      if (size != -1) return size;
-
-      size = 0;
-      {
-        int dataSize = 0;
-        for (int i = 0; i < stringListValue_.size(); i++) {
-          dataSize += computeStringSizeNoTag(stringListValue_.getRaw(i));
-        }
-        size += dataSize;
-        size += 1 * getStringListValueList().size();
-      }
-      if (!getDelimitedStringValueBytes().isEmpty()) {
-        size += com.google.protobuf.GeneratedMessageV3.computeStringSize(3, delimitedStringValue_);
-      }
-      size += unknownFields.getSerializedSize();
-      memoizedSize = size;
-      return size;
-    }
-
-    @java.lang.Override
-    public boolean equals(final java.lang.Object obj) {
-      if (obj == this) {
-       return true;
-      }
-      if (!(obj instanceof net.badata.protobuf.converter.proto.ResolverProto.ResolverTest)) {
-        return super.equals(obj);
-      }
-      net.badata.protobuf.converter.proto.ResolverProto.ResolverTest other = (net.badata.protobuf.converter.proto.ResolverProto.ResolverTest) obj;
-
-      boolean result = true;
-      result = result && getStringListValueList()
-          .equals(other.getStringListValueList());
-      result = result && getDelimitedStringValue()
-          .equals(other.getDelimitedStringValue());
-      result = result && unknownFields.equals(other.unknownFields);
-      return result;
-    }
-
-    @java.lang.Override
-    public int hashCode() {
-      if (memoizedHashCode != 0) {
-        return memoizedHashCode;
-      }
-      int hash = 41;
-      hash = (19 * hash) + getDescriptor().hashCode();
-      if (getStringListValueCount() > 0) {
-        hash = (37 * hash) + STRINGLISTVALUE_FIELD_NUMBER;
-        hash = (53 * hash) + getStringListValueList().hashCode();
-      }
-      hash = (37 * hash) + DELIMITEDSTRINGVALUE_FIELD_NUMBER;
-      hash = (53 * hash) + getDelimitedStringValue().hashCode();
-      hash = (29 * hash) + unknownFields.hashCode();
-      memoizedHashCode = hash;
-      return hash;
-    }
-
-    public static net.badata.protobuf.converter.proto.ResolverProto.ResolverTest parseFrom(
-        java.nio.ByteBuffer data)
-        throws com.google.protobuf.InvalidProtocolBufferException {
-      return PARSER.parseFrom(data);
-    }
-    public static net.badata.protobuf.converter.proto.ResolverProto.ResolverTest parseFrom(
-        java.nio.ByteBuffer data,
-        com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-        throws com.google.protobuf.InvalidProtocolBufferException {
-      return PARSER.parseFrom(data, extensionRegistry);
-    }
-    public static net.badata.protobuf.converter.proto.ResolverProto.ResolverTest parseFrom(
-        com.google.protobuf.ByteString data)
-        throws com.google.protobuf.InvalidProtocolBufferException {
-      return PARSER.parseFrom(data);
-    }
-    public static net.badata.protobuf.converter.proto.ResolverProto.ResolverTest parseFrom(
-        com.google.protobuf.ByteString data,
-        com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-        throws com.google.protobuf.InvalidProtocolBufferException {
-      return PARSER.parseFrom(data, extensionRegistry);
-    }
-    public static net.badata.protobuf.converter.proto.ResolverProto.ResolverTest parseFrom(byte[] data)
-        throws com.google.protobuf.InvalidProtocolBufferException {
-      return PARSER.parseFrom(data);
-    }
-    public static net.badata.protobuf.converter.proto.ResolverProto.ResolverTest parseFrom(
-        byte[] data,
-        com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-        throws com.google.protobuf.InvalidProtocolBufferException {
-      return PARSER.parseFrom(data, extensionRegistry);
-    }
-    public static net.badata.protobuf.converter.proto.ResolverProto.ResolverTest parseFrom(java.io.InputStream input)
-        throws java.io.IOException {
-      return com.google.protobuf.GeneratedMessageV3
-          .parseWithIOException(PARSER, input);
-    }
-    public static net.badata.protobuf.converter.proto.ResolverProto.ResolverTest parseFrom(
-        java.io.InputStream input,
-        com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-        throws java.io.IOException {
-      return com.google.protobuf.GeneratedMessageV3
-          .parseWithIOException(PARSER, input, extensionRegistry);
-    }
-    public static net.badata.protobuf.converter.proto.ResolverProto.ResolverTest parseDelimitedFrom(java.io.InputStream input)
-        throws java.io.IOException {
-      return com.google.protobuf.GeneratedMessageV3
-          .parseDelimitedWithIOException(PARSER, input);
-    }
-    public static net.badata.protobuf.converter.proto.ResolverProto.ResolverTest parseDelimitedFrom(
-        java.io.InputStream input,
-        com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-        throws java.io.IOException {
-      return com.google.protobuf.GeneratedMessageV3
-          .parseDelimitedWithIOException(PARSER, input, extensionRegistry);
-    }
-    public static net.badata.protobuf.converter.proto.ResolverProto.ResolverTest parseFrom(
-        com.google.protobuf.CodedInputStream input)
-        throws java.io.IOException {
-      return com.google.protobuf.GeneratedMessageV3
-          .parseWithIOException(PARSER, input);
-    }
-    public static net.badata.protobuf.converter.proto.ResolverProto.ResolverTest parseFrom(
-        com.google.protobuf.CodedInputStream input,
-        com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-        throws java.io.IOException {
-      return com.google.protobuf.GeneratedMessageV3
-          .parseWithIOException(PARSER, input, extensionRegistry);
-    }
-
-    public Builder newBuilderForType() { return newBuilder(); }
-    public static Builder newBuilder() {
-      return DEFAULT_INSTANCE.toBuilder();
-    }
-    public static Builder newBuilder(net.badata.protobuf.converter.proto.ResolverProto.ResolverTest prototype) {
-      return DEFAULT_INSTANCE.toBuilder().mergeFrom(prototype);
-    }
-    public Builder toBuilder() {
-      return this == DEFAULT_INSTANCE
-          ? new Builder() : new Builder().mergeFrom(this);
-    }
-
-    @java.lang.Override
-    protected Builder newBuilderForType(
-        com.google.protobuf.GeneratedMessageV3.BuilderParent parent) {
-      Builder builder = new Builder(parent);
-      return builder;
-    }
     /**
      * Protobuf type {@code net.badata.protobuf.converter.proto.ResolverTest}
      */
-    public static final class Builder extends
-        com.google.protobuf.GeneratedMessageV3.Builder<Builder> implements
-        // @@protoc_insertion_point(builder_implements:net.badata.protobuf.converter.proto.ResolverTest)
-        net.badata.protobuf.converter.proto.ResolverProto.ResolverTestOrBuilder {
-      public static final com.google.protobuf.Descriptors.Descriptor
-          getDescriptor() {
-        return net.badata.protobuf.converter.proto.ResolverProto.internal_static_net_badata_protobuf_converter_proto_ResolverTest_descriptor;
-      }
+    public static final class ResolverTest
+        extends com.google.protobuf.GeneratedMessageV3
+        implements
+            // @@protoc_insertion_point(message_implements:net.badata.protobuf.converter.proto.ResolverTest)
+            ResolverTestOrBuilder {
 
-      protected com.google.protobuf.GeneratedMessageV3.FieldAccessorTable
-          internalGetFieldAccessorTable() {
-        return net.badata.protobuf.converter.proto.ResolverProto.internal_static_net_badata_protobuf_converter_proto_ResolverTest_fieldAccessorTable
-            .ensureFieldAccessorsInitialized(
-                net.badata.protobuf.converter.proto.ResolverProto.ResolverTest.class, net.badata.protobuf.converter.proto.ResolverProto.ResolverTest.Builder.class);
-      }
+        private static final long serialVersionUID = 0L;
 
-      // Construct using net.badata.protobuf.converter.proto.ResolverProto.ResolverTest.newBuilder()
-      private Builder() {
-        maybeForceBuilderInitialization();
-      }
-
-      private Builder(
-          com.google.protobuf.GeneratedMessageV3.BuilderParent parent) {
-        super(parent);
-        maybeForceBuilderInitialization();
-      }
-      private void maybeForceBuilderInitialization() {
-        if (com.google.protobuf.GeneratedMessageV3
-                .alwaysUseFieldBuilders) {
+        // Use ResolverTest.newBuilder() to construct.
+        private ResolverTest(com.google.protobuf.GeneratedMessageV3.Builder<?> builder) {
+            super(builder);
         }
-      }
-      public Builder clear() {
-        super.clear();
-        stringListValue_ = com.google.protobuf.LazyStringArrayList.EMPTY;
-        bitField0_ = (bitField0_ & ~0x00000001);
-        delimitedStringValue_ = "";
 
-        return this;
-      }
-
-      public com.google.protobuf.Descriptors.Descriptor
-          getDescriptorForType() {
-        return net.badata.protobuf.converter.proto.ResolverProto.internal_static_net_badata_protobuf_converter_proto_ResolverTest_descriptor;
-      }
-
-      public net.badata.protobuf.converter.proto.ResolverProto.ResolverTest getDefaultInstanceForType() {
-        return net.badata.protobuf.converter.proto.ResolverProto.ResolverTest.getDefaultInstance();
-      }
-
-      public net.badata.protobuf.converter.proto.ResolverProto.ResolverTest build() {
-        net.badata.protobuf.converter.proto.ResolverProto.ResolverTest result = buildPartial();
-        if (!result.isInitialized()) {
-          throw newUninitializedMessageException(result);
+        private ResolverTest() {
+            stringListValue_ = com.google.protobuf.LazyStringArrayList.EMPTY;
+            delimitedStringValue_ = "";
         }
-        return result;
-      }
 
-      public net.badata.protobuf.converter.proto.ResolverProto.ResolverTest buildPartial() {
-        net.badata.protobuf.converter.proto.ResolverProto.ResolverTest result = new net.badata.protobuf.converter.proto.ResolverProto.ResolverTest(this);
-        int from_bitField0_ = bitField0_;
-        int to_bitField0_ = 0;
-        if (((bitField0_ & 0x00000001) == 0x00000001)) {
-          stringListValue_ = stringListValue_.getUnmodifiableView();
-          bitField0_ = (bitField0_ & ~0x00000001);
+        @java.lang.Override
+        public final com.google.protobuf.UnknownFieldSet getUnknownFields() {
+            return this.unknownFields;
         }
-        result.stringListValue_ = stringListValue_;
-        result.delimitedStringValue_ = delimitedStringValue_;
-        result.bitField0_ = to_bitField0_;
-        onBuilt();
-        return result;
-      }
 
-      public Builder clone() {
-        return (Builder) super.clone();
-      }
-      public Builder setField(
-          com.google.protobuf.Descriptors.FieldDescriptor field,
-          java.lang.Object value) {
-        return (Builder) super.setField(field, value);
-      }
-      public Builder clearField(
-          com.google.protobuf.Descriptors.FieldDescriptor field) {
-        return (Builder) super.clearField(field);
-      }
-      public Builder clearOneof(
-          com.google.protobuf.Descriptors.OneofDescriptor oneof) {
-        return (Builder) super.clearOneof(oneof);
-      }
-      public Builder setRepeatedField(
-          com.google.protobuf.Descriptors.FieldDescriptor field,
-          int index, java.lang.Object value) {
-        return (Builder) super.setRepeatedField(field, index, value);
-      }
-      public Builder addRepeatedField(
-          com.google.protobuf.Descriptors.FieldDescriptor field,
-          java.lang.Object value) {
-        return (Builder) super.addRepeatedField(field, value);
-      }
-      public Builder mergeFrom(com.google.protobuf.Message other) {
-        if (other instanceof net.badata.protobuf.converter.proto.ResolverProto.ResolverTest) {
-          return mergeFrom((net.badata.protobuf.converter.proto.ResolverProto.ResolverTest)other);
-        } else {
-          super.mergeFrom(other);
-          return this;
+        private ResolverTest(
+            com.google.protobuf.CodedInputStream input,
+            com.google.protobuf.ExtensionRegistryLite extensionRegistry
+        ) throws com.google.protobuf.InvalidProtocolBufferException {
+            this();
+            if (extensionRegistry == null) {
+                throw new java.lang.NullPointerException();
+            }
+            int mutable_bitField0_ = 0;
+            com.google.protobuf.UnknownFieldSet.Builder unknownFields = com.google.protobuf.UnknownFieldSet.newBuilder();
+            try {
+                boolean done = false;
+                while (!done) {
+                    int tag = input.readTag();
+                    switch (tag) {
+                        case 0:
+                            done = true;
+                            break;
+                        default:
+                            {
+                                if (!parseUnknownFieldProto3(input, unknownFields, extensionRegistry, tag)) {
+                                    done = true;
+                                }
+                                break;
+                            }
+                        case 10:
+                            {
+                                java.lang.String s = input.readStringRequireUtf8();
+                                if (!((mutable_bitField0_ & 0x00000001) == 0x00000001)) {
+                                    stringListValue_ = new com.google.protobuf.LazyStringArrayList();
+                                    mutable_bitField0_ |= 0x00000001;
+                                }
+                                stringListValue_.add(s);
+                                break;
+                            }
+                        case 26:
+                            {
+                                java.lang.String s = input.readStringRequireUtf8();
+
+                                delimitedStringValue_ = s;
+                                break;
+                            }
+                    }
+                }
+            } catch (com.google.protobuf.InvalidProtocolBufferException e) {
+                throw e.setUnfinishedMessage(this);
+            } catch (java.io.IOException e) {
+                throw new com.google.protobuf.InvalidProtocolBufferException(e).setUnfinishedMessage(this);
+            } finally {
+                if (((mutable_bitField0_ & 0x00000001) == 0x00000001)) {
+                    stringListValue_ = stringListValue_.getUnmodifiableView();
+                }
+                this.unknownFields = unknownFields.build();
+                makeExtensionsImmutable();
+            }
         }
-      }
 
-      public Builder mergeFrom(net.badata.protobuf.converter.proto.ResolverProto.ResolverTest other) {
-        if (other == net.badata.protobuf.converter.proto.ResolverProto.ResolverTest.getDefaultInstance()) return this;
-        if (!other.stringListValue_.isEmpty()) {
-          if (stringListValue_.isEmpty()) {
-            stringListValue_ = other.stringListValue_;
-            bitField0_ = (bitField0_ & ~0x00000001);
-          } else {
-            ensureStringListValueIsMutable();
-            stringListValue_.addAll(other.stringListValue_);
-          }
-          onChanged();
+        public static final com.google.protobuf.Descriptors.Descriptor getDescriptor() {
+            return net.badata.protobuf.converter.proto.ResolverProto.internal_static_net_badata_protobuf_converter_proto_ResolverTest_descriptor;
         }
-        if (!other.getDelimitedStringValue().isEmpty()) {
-          delimitedStringValue_ = other.delimitedStringValue_;
-          onChanged();
+
+        protected com.google.protobuf.GeneratedMessageV3.FieldAccessorTable internalGetFieldAccessorTable() {
+            return net.badata.protobuf.converter.proto.ResolverProto.internal_static_net_badata_protobuf_converter_proto_ResolverTest_fieldAccessorTable.ensureFieldAccessorsInitialized(
+                net.badata.protobuf.converter.proto.ResolverProto.ResolverTest.class,
+                net.badata.protobuf.converter.proto.ResolverProto.ResolverTest.Builder.class
+            );
         }
-        this.mergeUnknownFields(other.unknownFields);
-        onChanged();
-        return this;
-      }
 
-      public final boolean isInitialized() {
-        return true;
-      }
+        private int bitField0_;
+        public static final int STRINGLISTVALUE_FIELD_NUMBER = 1;
+        private com.google.protobuf.LazyStringList stringListValue_;
 
-      public Builder mergeFrom(
-          com.google.protobuf.CodedInputStream input,
-          com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-          throws java.io.IOException {
-        net.badata.protobuf.converter.proto.ResolverProto.ResolverTest parsedMessage = null;
-        try {
-          parsedMessage = PARSER.parsePartialFrom(input, extensionRegistry);
-        } catch (com.google.protobuf.InvalidProtocolBufferException e) {
-          parsedMessage = (net.badata.protobuf.converter.proto.ResolverProto.ResolverTest) e.getUnfinishedMessage();
-          throw e.unwrapIOException();
-        } finally {
-          if (parsedMessage != null) {
-            mergeFrom(parsedMessage);
-          }
+        /**
+         * <code>repeated string stringListValue = 1;</code>
+         */
+        public com.google.protobuf.ProtocolStringList getStringListValueList() {
+            return stringListValue_;
         }
-        return this;
-      }
-      private int bitField0_;
 
-      private com.google.protobuf.LazyStringList stringListValue_ = com.google.protobuf.LazyStringArrayList.EMPTY;
-      private void ensureStringListValueIsMutable() {
-        if (!((bitField0_ & 0x00000001) == 0x00000001)) {
-          stringListValue_ = new com.google.protobuf.LazyStringArrayList(stringListValue_);
-          bitField0_ |= 0x00000001;
-         }
-      }
-      /**
-       * <code>repeated string stringListValue = 1;</code>
-       */
-      public com.google.protobuf.ProtocolStringList
-          getStringListValueList() {
-        return stringListValue_.getUnmodifiableView();
-      }
-      /**
-       * <code>repeated string stringListValue = 1;</code>
-       */
-      public int getStringListValueCount() {
-        return stringListValue_.size();
-      }
-      /**
-       * <code>repeated string stringListValue = 1;</code>
-       */
-      public java.lang.String getStringListValue(int index) {
-        return stringListValue_.get(index);
-      }
-      /**
-       * <code>repeated string stringListValue = 1;</code>
-       */
-      public com.google.protobuf.ByteString
-          getStringListValueBytes(int index) {
-        return stringListValue_.getByteString(index);
-      }
-      /**
-       * <code>repeated string stringListValue = 1;</code>
-       */
-      public Builder setStringListValue(
-          int index, java.lang.String value) {
-        if (value == null) {
-    throw new NullPointerException();
-  }
-  ensureStringListValueIsMutable();
-        stringListValue_.set(index, value);
-        onChanged();
-        return this;
-      }
-      /**
-       * <code>repeated string stringListValue = 1;</code>
-       */
-      public Builder addStringListValue(
-          java.lang.String value) {
-        if (value == null) {
-    throw new NullPointerException();
-  }
-  ensureStringListValueIsMutable();
-        stringListValue_.add(value);
-        onChanged();
-        return this;
-      }
-      /**
-       * <code>repeated string stringListValue = 1;</code>
-       */
-      public Builder addAllStringListValue(
-          java.lang.Iterable<java.lang.String> values) {
-        ensureStringListValueIsMutable();
-        com.google.protobuf.AbstractMessageLite.Builder.addAll(
-            values, stringListValue_);
-        onChanged();
-        return this;
-      }
-      /**
-       * <code>repeated string stringListValue = 1;</code>
-       */
-      public Builder clearStringListValue() {
-        stringListValue_ = com.google.protobuf.LazyStringArrayList.EMPTY;
-        bitField0_ = (bitField0_ & ~0x00000001);
-        onChanged();
-        return this;
-      }
-      /**
-       * <code>repeated string stringListValue = 1;</code>
-       */
-      public Builder addStringListValueBytes(
-          com.google.protobuf.ByteString value) {
-        if (value == null) {
-    throw new NullPointerException();
-  }
-  checkByteStringIsUtf8(value);
-        ensureStringListValueIsMutable();
-        stringListValue_.add(value);
-        onChanged();
-        return this;
-      }
-
-      private java.lang.Object delimitedStringValue_ = "";
-      /**
-       * <code>string delimitedStringValue = 3;</code>
-       */
-      public java.lang.String getDelimitedStringValue() {
-        java.lang.Object ref = delimitedStringValue_;
-        if (!(ref instanceof java.lang.String)) {
-          com.google.protobuf.ByteString bs =
-              (com.google.protobuf.ByteString) ref;
-          java.lang.String s = bs.toStringUtf8();
-          delimitedStringValue_ = s;
-          return s;
-        } else {
-          return (java.lang.String) ref;
+        /**
+         * <code>repeated string stringListValue = 1;</code>
+         */
+        public int getStringListValueCount() {
+            return stringListValue_.size();
         }
-      }
-      /**
-       * <code>string delimitedStringValue = 3;</code>
-       */
-      public com.google.protobuf.ByteString
-          getDelimitedStringValueBytes() {
-        java.lang.Object ref = delimitedStringValue_;
-        if (ref instanceof String) {
-          com.google.protobuf.ByteString b = 
-              com.google.protobuf.ByteString.copyFromUtf8(
-                  (java.lang.String) ref);
-          delimitedStringValue_ = b;
-          return b;
-        } else {
-          return (com.google.protobuf.ByteString) ref;
+
+        /**
+         * <code>repeated string stringListValue = 1;</code>
+         */
+        public java.lang.String getStringListValue(int index) {
+            return stringListValue_.get(index);
         }
-      }
-      /**
-       * <code>string delimitedStringValue = 3;</code>
-       */
-      public Builder setDelimitedStringValue(
-          java.lang.String value) {
-        if (value == null) {
-    throw new NullPointerException();
-  }
-  
-        delimitedStringValue_ = value;
-        onChanged();
-        return this;
-      }
-      /**
-       * <code>string delimitedStringValue = 3;</code>
-       */
-      public Builder clearDelimitedStringValue() {
-        
-        delimitedStringValue_ = getDefaultInstance().getDelimitedStringValue();
-        onChanged();
-        return this;
-      }
-      /**
-       * <code>string delimitedStringValue = 3;</code>
-       */
-      public Builder setDelimitedStringValueBytes(
-          com.google.protobuf.ByteString value) {
-        if (value == null) {
-    throw new NullPointerException();
-  }
-  checkByteStringIsUtf8(value);
-        
-        delimitedStringValue_ = value;
-        onChanged();
-        return this;
-      }
-      public final Builder setUnknownFields(
-          final com.google.protobuf.UnknownFieldSet unknownFields) {
-        return super.setUnknownFieldsProto3(unknownFields);
-      }
 
-      public final Builder mergeUnknownFields(
-          final com.google.protobuf.UnknownFieldSet unknownFields) {
-        return super.mergeUnknownFields(unknownFields);
-      }
+        /**
+         * <code>repeated string stringListValue = 1;</code>
+         */
+        public com.google.protobuf.ByteString getStringListValueBytes(int index) {
+            return stringListValue_.getByteString(index);
+        }
 
+        public static final int DELIMITEDSTRINGVALUE_FIELD_NUMBER = 3;
+        private volatile java.lang.Object delimitedStringValue_;
 
-      // @@protoc_insertion_point(builder_scope:net.badata.protobuf.converter.proto.ResolverTest)
-    }
+        /**
+         * <code>string delimitedStringValue = 3;</code>
+         */
+        public java.lang.String getDelimitedStringValue() {
+            java.lang.Object ref = delimitedStringValue_;
+            if (ref instanceof java.lang.String) {
+                return (java.lang.String) ref;
+            } else {
+                com.google.protobuf.ByteString bs = (com.google.protobuf.ByteString) ref;
+                java.lang.String s = bs.toStringUtf8();
+                delimitedStringValue_ = s;
+                return s;
+            }
+        }
 
-    // @@protoc_insertion_point(class_scope:net.badata.protobuf.converter.proto.ResolverTest)
-    private static final net.badata.protobuf.converter.proto.ResolverProto.ResolverTest DEFAULT_INSTANCE;
-    static {
-      DEFAULT_INSTANCE = new net.badata.protobuf.converter.proto.ResolverProto.ResolverTest();
-    }
+        /**
+         * <code>string delimitedStringValue = 3;</code>
+         */
+        public com.google.protobuf.ByteString getDelimitedStringValueBytes() {
+            java.lang.Object ref = delimitedStringValue_;
+            if (ref instanceof java.lang.String) {
+                com.google.protobuf.ByteString b = com.google.protobuf.ByteString.copyFromUtf8((java.lang.String) ref);
+                delimitedStringValue_ = b;
+                return b;
+            } else {
+                return (com.google.protobuf.ByteString) ref;
+            }
+        }
 
-    public static net.badata.protobuf.converter.proto.ResolverProto.ResolverTest getDefaultInstance() {
-      return DEFAULT_INSTANCE;
-    }
+        private byte memoizedIsInitialized = -1;
 
-    private static final com.google.protobuf.Parser<ResolverTest>
-        PARSER = new com.google.protobuf.AbstractParser<ResolverTest>() {
-      public ResolverTest parsePartialFrom(
-          com.google.protobuf.CodedInputStream input,
-          com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-          throws com.google.protobuf.InvalidProtocolBufferException {
-        return new ResolverTest(input, extensionRegistry);
-      }
-    };
+        public final boolean isInitialized() {
+            byte isInitialized = memoizedIsInitialized;
+            if (isInitialized == 1) return true;
+            if (isInitialized == 0) return false;
 
-    public static com.google.protobuf.Parser<ResolverTest> parser() {
-      return PARSER;
-    }
+            memoizedIsInitialized = 1;
+            return true;
+        }
 
-    @java.lang.Override
-    public com.google.protobuf.Parser<ResolverTest> getParserForType() {
-      return PARSER;
-    }
+        public void writeTo(com.google.protobuf.CodedOutputStream output) throws java.io.IOException {
+            for (int i = 0; i < stringListValue_.size(); i++) {
+                com.google.protobuf.GeneratedMessageV3.writeString(output, 1, stringListValue_.getRaw(i));
+            }
+            if (!getDelimitedStringValueBytes().isEmpty()) {
+                com.google.protobuf.GeneratedMessageV3.writeString(output, 3, delimitedStringValue_);
+            }
+            unknownFields.writeTo(output);
+        }
 
-    public net.badata.protobuf.converter.proto.ResolverProto.ResolverTest getDefaultInstanceForType() {
-      return DEFAULT_INSTANCE;
-    }
+        public int getSerializedSize() {
+            int size = memoizedSize;
+            if (size != -1) return size;
 
-  }
+            size = 0;
+            {
+                int dataSize = 0;
+                for (int i = 0; i < stringListValue_.size(); i++) {
+                    dataSize += computeStringSizeNoTag(stringListValue_.getRaw(i));
+                }
+                size += dataSize;
+                size += 1 * getStringListValueList().size();
+            }
+            if (!getDelimitedStringValueBytes().isEmpty()) {
+                size += com.google.protobuf.GeneratedMessageV3.computeStringSize(3, delimitedStringValue_);
+            }
+            size += unknownFields.getSerializedSize();
+            memoizedSize = size;
+            return size;
+        }
 
-  private static final com.google.protobuf.Descriptors.Descriptor
-    internal_static_net_badata_protobuf_converter_proto_ResolverTest_descriptor;
-  private static final 
-    com.google.protobuf.GeneratedMessageV3.FieldAccessorTable
-      internal_static_net_badata_protobuf_converter_proto_ResolverTest_fieldAccessorTable;
+        @java.lang.Override
+        public boolean equals(final java.lang.Object obj) {
+            if (obj == this) {
+                return true;
+            }
+            if (!(obj instanceof net.badata.protobuf.converter.proto.ResolverProto.ResolverTest)) {
+                return super.equals(obj);
+            }
+            net.badata.protobuf.converter.proto.ResolverProto.ResolverTest other = (net.badata.protobuf.converter.proto.ResolverProto.ResolverTest) obj;
 
-  public static com.google.protobuf.Descriptors.FileDescriptor
-      getDescriptor() {
-    return descriptor;
-  }
-  private static  com.google.protobuf.Descriptors.FileDescriptor
-      descriptor;
-  static {
-    java.lang.String[] descriptorData = {
-      "\n\031field_resolver_test.proto\022#net.badata." +
-      "protobuf.converter.proto\"E\n\014ResolverTest" +
-      "\022\027\n\017stringListValue\030\001 \003(\t\022\034\n\024delimitedSt" +
-      "ringValue\030\003 \001(\tB4\n#net.badata.protobuf.c" +
-      "onverter.protoB\rResolverProtob\006proto3"
-    };
-    com.google.protobuf.Descriptors.FileDescriptor.InternalDescriptorAssigner assigner =
-        new com.google.protobuf.Descriptors.FileDescriptor.    InternalDescriptorAssigner() {
-          public com.google.protobuf.ExtensionRegistry assignDescriptors(
-              com.google.protobuf.Descriptors.FileDescriptor root) {
-            descriptor = root;
-            return null;
-          }
+            boolean result = true;
+            result = result && getStringListValueList().equals(other.getStringListValueList());
+            result = result && getDelimitedStringValue().equals(other.getDelimitedStringValue());
+            result = result && unknownFields.equals(other.unknownFields);
+            return result;
+        }
+
+        @java.lang.Override
+        public int hashCode() {
+            if (memoizedHashCode != 0) {
+                return memoizedHashCode;
+            }
+            int hash = 41;
+            hash = (19 * hash) + getDescriptor().hashCode();
+            if (getStringListValueCount() > 0) {
+                hash = (37 * hash) + STRINGLISTVALUE_FIELD_NUMBER;
+                hash = (53 * hash) + getStringListValueList().hashCode();
+            }
+            hash = (37 * hash) + DELIMITEDSTRINGVALUE_FIELD_NUMBER;
+            hash = (53 * hash) + getDelimitedStringValue().hashCode();
+            hash = (29 * hash) + unknownFields.hashCode();
+            memoizedHashCode = hash;
+            return hash;
+        }
+
+        public static net.badata.protobuf.converter.proto.ResolverProto.ResolverTest parseFrom(
+            java.nio.ByteBuffer data
+        ) throws com.google.protobuf.InvalidProtocolBufferException {
+            return PARSER.parseFrom(data);
+        }
+
+        public static net.badata.protobuf.converter.proto.ResolverProto.ResolverTest parseFrom(
+            java.nio.ByteBuffer data,
+            com.google.protobuf.ExtensionRegistryLite extensionRegistry
+        ) throws com.google.protobuf.InvalidProtocolBufferException {
+            return PARSER.parseFrom(data, extensionRegistry);
+        }
+
+        public static net.badata.protobuf.converter.proto.ResolverProto.ResolverTest parseFrom(
+            com.google.protobuf.ByteString data
+        ) throws com.google.protobuf.InvalidProtocolBufferException {
+            return PARSER.parseFrom(data);
+        }
+
+        public static net.badata.protobuf.converter.proto.ResolverProto.ResolverTest parseFrom(
+            com.google.protobuf.ByteString data,
+            com.google.protobuf.ExtensionRegistryLite extensionRegistry
+        ) throws com.google.protobuf.InvalidProtocolBufferException {
+            return PARSER.parseFrom(data, extensionRegistry);
+        }
+
+        public static net.badata.protobuf.converter.proto.ResolverProto.ResolverTest parseFrom(byte[] data)
+            throws com.google.protobuf.InvalidProtocolBufferException {
+            return PARSER.parseFrom(data);
+        }
+
+        public static net.badata.protobuf.converter.proto.ResolverProto.ResolverTest parseFrom(
+            byte[] data,
+            com.google.protobuf.ExtensionRegistryLite extensionRegistry
+        ) throws com.google.protobuf.InvalidProtocolBufferException {
+            return PARSER.parseFrom(data, extensionRegistry);
+        }
+
+        public static net.badata.protobuf.converter.proto.ResolverProto.ResolverTest parseFrom(
+            java.io.InputStream input
+        ) throws java.io.IOException {
+            return com.google.protobuf.GeneratedMessageV3.parseWithIOException(PARSER, input);
+        }
+
+        public static net.badata.protobuf.converter.proto.ResolverProto.ResolverTest parseFrom(
+            java.io.InputStream input,
+            com.google.protobuf.ExtensionRegistryLite extensionRegistry
+        ) throws java.io.IOException {
+            return com.google.protobuf.GeneratedMessageV3.parseWithIOException(PARSER, input, extensionRegistry);
+        }
+
+        public static net.badata.protobuf.converter.proto.ResolverProto.ResolverTest parseDelimitedFrom(
+            java.io.InputStream input
+        ) throws java.io.IOException {
+            return com.google.protobuf.GeneratedMessageV3.parseDelimitedWithIOException(PARSER, input);
+        }
+
+        public static net.badata.protobuf.converter.proto.ResolverProto.ResolverTest parseDelimitedFrom(
+            java.io.InputStream input,
+            com.google.protobuf.ExtensionRegistryLite extensionRegistry
+        ) throws java.io.IOException {
+            return com.google.protobuf.GeneratedMessageV3.parseDelimitedWithIOException(
+                PARSER,
+                input,
+                extensionRegistry
+            );
+        }
+
+        public static net.badata.protobuf.converter.proto.ResolverProto.ResolverTest parseFrom(
+            com.google.protobuf.CodedInputStream input
+        ) throws java.io.IOException {
+            return com.google.protobuf.GeneratedMessageV3.parseWithIOException(PARSER, input);
+        }
+
+        public static net.badata.protobuf.converter.proto.ResolverProto.ResolverTest parseFrom(
+            com.google.protobuf.CodedInputStream input,
+            com.google.protobuf.ExtensionRegistryLite extensionRegistry
+        ) throws java.io.IOException {
+            return com.google.protobuf.GeneratedMessageV3.parseWithIOException(PARSER, input, extensionRegistry);
+        }
+
+        public Builder newBuilderForType() {
+            return newBuilder();
+        }
+
+        public static Builder newBuilder() {
+            return DEFAULT_INSTANCE.toBuilder();
+        }
+
+        public static Builder newBuilder(net.badata.protobuf.converter.proto.ResolverProto.ResolverTest prototype) {
+            return DEFAULT_INSTANCE.toBuilder().mergeFrom(prototype);
+        }
+
+        public Builder toBuilder() {
+            return this == DEFAULT_INSTANCE ? new Builder() : new Builder().mergeFrom(this);
+        }
+
+        @java.lang.Override
+        protected Builder newBuilderForType(com.google.protobuf.GeneratedMessageV3.BuilderParent parent) {
+            Builder builder = new Builder(parent);
+            return builder;
+        }
+
+        /**
+         * Protobuf type {@code net.badata.protobuf.converter.proto.ResolverTest}
+         */
+        public static final class Builder
+            extends com.google.protobuf.GeneratedMessageV3.Builder<Builder>
+            implements
+                // @@protoc_insertion_point(builder_implements:net.badata.protobuf.converter.proto.ResolverTest)
+                net.badata.protobuf.converter.proto.ResolverProto.ResolverTestOrBuilder {
+
+            public static final com.google.protobuf.Descriptors.Descriptor getDescriptor() {
+                return net.badata.protobuf.converter.proto.ResolverProto.internal_static_net_badata_protobuf_converter_proto_ResolverTest_descriptor;
+            }
+
+            protected com.google.protobuf.GeneratedMessageV3.FieldAccessorTable internalGetFieldAccessorTable() {
+                return net.badata.protobuf.converter.proto.ResolverProto.internal_static_net_badata_protobuf_converter_proto_ResolverTest_fieldAccessorTable.ensureFieldAccessorsInitialized(
+                    net.badata.protobuf.converter.proto.ResolverProto.ResolverTest.class,
+                    net.badata.protobuf.converter.proto.ResolverProto.ResolverTest.Builder.class
+                );
+            }
+
+            // Construct using net.badata.protobuf.converter.proto.ResolverProto.ResolverTest.newBuilder()
+            private Builder() {
+                maybeForceBuilderInitialization();
+            }
+
+            private Builder(com.google.protobuf.GeneratedMessageV3.BuilderParent parent) {
+                super(parent);
+                maybeForceBuilderInitialization();
+            }
+
+            private void maybeForceBuilderInitialization() {
+                if (com.google.protobuf.GeneratedMessageV3.alwaysUseFieldBuilders) {}
+            }
+
+            public Builder clear() {
+                super.clear();
+                stringListValue_ = com.google.protobuf.LazyStringArrayList.EMPTY;
+                bitField0_ = (bitField0_ & ~0x00000001);
+                delimitedStringValue_ = "";
+
+                return this;
+            }
+
+            public com.google.protobuf.Descriptors.Descriptor getDescriptorForType() {
+                return net.badata.protobuf.converter.proto.ResolverProto.internal_static_net_badata_protobuf_converter_proto_ResolverTest_descriptor;
+            }
+
+            public net.badata.protobuf.converter.proto.ResolverProto.ResolverTest getDefaultInstanceForType() {
+                return net.badata.protobuf.converter.proto.ResolverProto.ResolverTest.getDefaultInstance();
+            }
+
+            public net.badata.protobuf.converter.proto.ResolverProto.ResolverTest build() {
+                net.badata.protobuf.converter.proto.ResolverProto.ResolverTest result = buildPartial();
+                if (!result.isInitialized()) {
+                    throw newUninitializedMessageException(result);
+                }
+                return result;
+            }
+
+            public net.badata.protobuf.converter.proto.ResolverProto.ResolverTest buildPartial() {
+                net.badata.protobuf.converter.proto.ResolverProto.ResolverTest result = new net.badata.protobuf.converter.proto.ResolverProto.ResolverTest(
+                    this
+                );
+                int from_bitField0_ = bitField0_;
+                int to_bitField0_ = 0;
+                if (((bitField0_ & 0x00000001) == 0x00000001)) {
+                    stringListValue_ = stringListValue_.getUnmodifiableView();
+                    bitField0_ = (bitField0_ & ~0x00000001);
+                }
+                result.stringListValue_ = stringListValue_;
+                result.delimitedStringValue_ = delimitedStringValue_;
+                result.bitField0_ = to_bitField0_;
+                onBuilt();
+                return result;
+            }
+
+            public Builder clone() {
+                return (Builder) super.clone();
+            }
+
+            public Builder setField(com.google.protobuf.Descriptors.FieldDescriptor field, java.lang.Object value) {
+                return (Builder) super.setField(field, value);
+            }
+
+            public Builder clearField(com.google.protobuf.Descriptors.FieldDescriptor field) {
+                return (Builder) super.clearField(field);
+            }
+
+            public Builder clearOneof(com.google.protobuf.Descriptors.OneofDescriptor oneof) {
+                return (Builder) super.clearOneof(oneof);
+            }
+
+            public Builder setRepeatedField(
+                com.google.protobuf.Descriptors.FieldDescriptor field,
+                int index,
+                java.lang.Object value
+            ) {
+                return (Builder) super.setRepeatedField(field, index, value);
+            }
+
+            public Builder addRepeatedField(
+                com.google.protobuf.Descriptors.FieldDescriptor field,
+                java.lang.Object value
+            ) {
+                return (Builder) super.addRepeatedField(field, value);
+            }
+
+            public Builder mergeFrom(com.google.protobuf.Message other) {
+                if (other instanceof net.badata.protobuf.converter.proto.ResolverProto.ResolverTest) {
+                    return mergeFrom((net.badata.protobuf.converter.proto.ResolverProto.ResolverTest) other);
+                } else {
+                    super.mergeFrom(other);
+                    return this;
+                }
+            }
+
+            public Builder mergeFrom(net.badata.protobuf.converter.proto.ResolverProto.ResolverTest other) {
+                if (
+                    other == net.badata.protobuf.converter.proto.ResolverProto.ResolverTest.getDefaultInstance()
+                ) return this;
+                if (!other.stringListValue_.isEmpty()) {
+                    if (stringListValue_.isEmpty()) {
+                        stringListValue_ = other.stringListValue_;
+                        bitField0_ = (bitField0_ & ~0x00000001);
+                    } else {
+                        ensureStringListValueIsMutable();
+                        stringListValue_.addAll(other.stringListValue_);
+                    }
+                    onChanged();
+                }
+                if (!other.getDelimitedStringValue().isEmpty()) {
+                    delimitedStringValue_ = other.delimitedStringValue_;
+                    onChanged();
+                }
+                this.mergeUnknownFields(other.unknownFields);
+                onChanged();
+                return this;
+            }
+
+            public final boolean isInitialized() {
+                return true;
+            }
+
+            public Builder mergeFrom(
+                com.google.protobuf.CodedInputStream input,
+                com.google.protobuf.ExtensionRegistryLite extensionRegistry
+            ) throws java.io.IOException {
+                net.badata.protobuf.converter.proto.ResolverProto.ResolverTest parsedMessage = null;
+                try {
+                    parsedMessage = PARSER.parsePartialFrom(input, extensionRegistry);
+                } catch (com.google.protobuf.InvalidProtocolBufferException e) {
+                    parsedMessage =
+                        (net.badata.protobuf.converter.proto.ResolverProto.ResolverTest) e.getUnfinishedMessage();
+                    throw e.unwrapIOException();
+                } finally {
+                    if (parsedMessage != null) {
+                        mergeFrom(parsedMessage);
+                    }
+                }
+                return this;
+            }
+
+            private int bitField0_;
+
+            private com.google.protobuf.LazyStringList stringListValue_ = com.google.protobuf.LazyStringArrayList.EMPTY;
+
+            private void ensureStringListValueIsMutable() {
+                if (!((bitField0_ & 0x00000001) == 0x00000001)) {
+                    stringListValue_ = new com.google.protobuf.LazyStringArrayList(stringListValue_);
+                    bitField0_ |= 0x00000001;
+                }
+            }
+
+            /**
+             * <code>repeated string stringListValue = 1;</code>
+             */
+            public com.google.protobuf.ProtocolStringList getStringListValueList() {
+                return stringListValue_.getUnmodifiableView();
+            }
+
+            /**
+             * <code>repeated string stringListValue = 1;</code>
+             */
+            public int getStringListValueCount() {
+                return stringListValue_.size();
+            }
+
+            /**
+             * <code>repeated string stringListValue = 1;</code>
+             */
+            public java.lang.String getStringListValue(int index) {
+                return stringListValue_.get(index);
+            }
+
+            /**
+             * <code>repeated string stringListValue = 1;</code>
+             */
+            public com.google.protobuf.ByteString getStringListValueBytes(int index) {
+                return stringListValue_.getByteString(index);
+            }
+
+            /**
+             * <code>repeated string stringListValue = 1;</code>
+             */
+            public Builder setStringListValue(int index, java.lang.String value) {
+                if (value == null) {
+                    throw new NullPointerException();
+                }
+                ensureStringListValueIsMutable();
+                stringListValue_.set(index, value);
+                onChanged();
+                return this;
+            }
+
+            /**
+             * <code>repeated string stringListValue = 1;</code>
+             */
+            public Builder addStringListValue(java.lang.String value) {
+                if (value == null) {
+                    throw new NullPointerException();
+                }
+                ensureStringListValueIsMutable();
+                stringListValue_.add(value);
+                onChanged();
+                return this;
+            }
+
+            /**
+             * <code>repeated string stringListValue = 1;</code>
+             */
+            public Builder addAllStringListValue(java.lang.Iterable<java.lang.String> values) {
+                ensureStringListValueIsMutable();
+                com.google.protobuf.AbstractMessageLite.Builder.addAll(values, stringListValue_);
+                onChanged();
+                return this;
+            }
+
+            /**
+             * <code>repeated string stringListValue = 1;</code>
+             */
+            public Builder clearStringListValue() {
+                stringListValue_ = com.google.protobuf.LazyStringArrayList.EMPTY;
+                bitField0_ = (bitField0_ & ~0x00000001);
+                onChanged();
+                return this;
+            }
+
+            /**
+             * <code>repeated string stringListValue = 1;</code>
+             */
+            public Builder addStringListValueBytes(com.google.protobuf.ByteString value) {
+                if (value == null) {
+                    throw new NullPointerException();
+                }
+                checkByteStringIsUtf8(value);
+                ensureStringListValueIsMutable();
+                stringListValue_.add(value);
+                onChanged();
+                return this;
+            }
+
+            private java.lang.Object delimitedStringValue_ = "";
+
+            /**
+             * <code>string delimitedStringValue = 3;</code>
+             */
+            public java.lang.String getDelimitedStringValue() {
+                java.lang.Object ref = delimitedStringValue_;
+                if (!(ref instanceof java.lang.String)) {
+                    com.google.protobuf.ByteString bs = (com.google.protobuf.ByteString) ref;
+                    java.lang.String s = bs.toStringUtf8();
+                    delimitedStringValue_ = s;
+                    return s;
+                } else {
+                    return (java.lang.String) ref;
+                }
+            }
+
+            /**
+             * <code>string delimitedStringValue = 3;</code>
+             */
+            public com.google.protobuf.ByteString getDelimitedStringValueBytes() {
+                java.lang.Object ref = delimitedStringValue_;
+                if (ref instanceof String) {
+                    com.google.protobuf.ByteString b = com.google.protobuf.ByteString.copyFromUtf8(
+                        (java.lang.String) ref
+                    );
+                    delimitedStringValue_ = b;
+                    return b;
+                } else {
+                    return (com.google.protobuf.ByteString) ref;
+                }
+            }
+
+            /**
+             * <code>string delimitedStringValue = 3;</code>
+             */
+            public Builder setDelimitedStringValue(java.lang.String value) {
+                if (value == null) {
+                    throw new NullPointerException();
+                }
+
+                delimitedStringValue_ = value;
+                onChanged();
+                return this;
+            }
+
+            /**
+             * <code>string delimitedStringValue = 3;</code>
+             */
+            public Builder clearDelimitedStringValue() {
+                delimitedStringValue_ = getDefaultInstance().getDelimitedStringValue();
+                onChanged();
+                return this;
+            }
+
+            /**
+             * <code>string delimitedStringValue = 3;</code>
+             */
+            public Builder setDelimitedStringValueBytes(com.google.protobuf.ByteString value) {
+                if (value == null) {
+                    throw new NullPointerException();
+                }
+                checkByteStringIsUtf8(value);
+
+                delimitedStringValue_ = value;
+                onChanged();
+                return this;
+            }
+
+            public final Builder setUnknownFields(final com.google.protobuf.UnknownFieldSet unknownFields) {
+                return super.setUnknownFieldsProto3(unknownFields);
+            }
+
+            public final Builder mergeUnknownFields(final com.google.protobuf.UnknownFieldSet unknownFields) {
+                return super.mergeUnknownFields(unknownFields);
+            }
+            // @@protoc_insertion_point(builder_scope:net.badata.protobuf.converter.proto.ResolverTest)
+        }
+
+        // @@protoc_insertion_point(class_scope:net.badata.protobuf.converter.proto.ResolverTest)
+        private static final net.badata.protobuf.converter.proto.ResolverProto.ResolverTest DEFAULT_INSTANCE;
+
+        static {
+            DEFAULT_INSTANCE = new net.badata.protobuf.converter.proto.ResolverProto.ResolverTest();
+        }
+
+        public static net.badata.protobuf.converter.proto.ResolverProto.ResolverTest getDefaultInstance() {
+            return DEFAULT_INSTANCE;
+        }
+
+        private static final com.google.protobuf.Parser<ResolverTest> PARSER = new com.google.protobuf.AbstractParser<ResolverTest>() {
+            public ResolverTest parsePartialFrom(
+                com.google.protobuf.CodedInputStream input,
+                com.google.protobuf.ExtensionRegistryLite extensionRegistry
+            ) throws com.google.protobuf.InvalidProtocolBufferException {
+                return new ResolverTest(input, extensionRegistry);
+            }
         };
-    com.google.protobuf.Descriptors.FileDescriptor
-      .internalBuildGeneratedFileFrom(descriptorData,
-        new com.google.protobuf.Descriptors.FileDescriptor[] {
-        }, assigner);
-    internal_static_net_badata_protobuf_converter_proto_ResolverTest_descriptor =
-      getDescriptor().getMessageTypes().get(0);
-    internal_static_net_badata_protobuf_converter_proto_ResolverTest_fieldAccessorTable = new
-      com.google.protobuf.GeneratedMessageV3.FieldAccessorTable(
-        internal_static_net_badata_protobuf_converter_proto_ResolverTest_descriptor,
-        new java.lang.String[] { "StringListValue", "DelimitedStringValue", });
-  }
 
-  // @@protoc_insertion_point(outer_class_scope)
+        public static com.google.protobuf.Parser<ResolverTest> parser() {
+            return PARSER;
+        }
+
+        @java.lang.Override
+        public com.google.protobuf.Parser<ResolverTest> getParserForType() {
+            return PARSER;
+        }
+
+        public net.badata.protobuf.converter.proto.ResolverProto.ResolverTest getDefaultInstanceForType() {
+            return DEFAULT_INSTANCE;
+        }
+    }
+
+    private static final com.google.protobuf.Descriptors.Descriptor internal_static_net_badata_protobuf_converter_proto_ResolverTest_descriptor;
+    private static final com.google.protobuf.GeneratedMessageV3.FieldAccessorTable internal_static_net_badata_protobuf_converter_proto_ResolverTest_fieldAccessorTable;
+
+    public static com.google.protobuf.Descriptors.FileDescriptor getDescriptor() {
+        return descriptor;
+    }
+
+    private static com.google.protobuf.Descriptors.FileDescriptor descriptor;
+
+    static {
+        java.lang.String[] descriptorData = {
+            "\n\031field_resolver_test.proto\022#net.badata." +
+            "protobuf.converter.proto\"E\n\014ResolverTest" +
+            "\022\027\n\017stringListValue\030\001 \003(\t\022\034\n\024delimitedSt" +
+            "ringValue\030\003 \001(\tB4\n#net.badata.protobuf.c" +
+            "onverter.protoB\rResolverProtob\006proto3",
+        };
+        com.google.protobuf.Descriptors.FileDescriptor.InternalDescriptorAssigner assigner = new com.google.protobuf.Descriptors.FileDescriptor.InternalDescriptorAssigner() {
+            public com.google.protobuf.ExtensionRegistry assignDescriptors(
+                com.google.protobuf.Descriptors.FileDescriptor root
+            ) {
+                descriptor = root;
+                return null;
+            }
+        };
+        com.google.protobuf.Descriptors.FileDescriptor.internalBuildGeneratedFileFrom(
+            descriptorData,
+            new com.google.protobuf.Descriptors.FileDescriptor[] {},
+            assigner
+        );
+        internal_static_net_badata_protobuf_converter_proto_ResolverTest_descriptor =
+            getDescriptor().getMessageTypes().get(0);
+        internal_static_net_badata_protobuf_converter_proto_ResolverTest_fieldAccessorTable =
+            new com.google.protobuf.GeneratedMessageV3.FieldAccessorTable(
+                internal_static_net_badata_protobuf_converter_proto_ResolverTest_descriptor,
+                new java.lang.String[] { "StringListValue", "DelimitedStringValue" }
+            );
+    }
+    // @@protoc_insertion_point(outer_class_scope)
 }
